Update et-uv-pro
When the jq (json query) command executes, it returns 0 if there is a .bluetooth.mac node in the json (which there is, and should be).  This means we can't use $? to see if there is a value in the node or not.  We need to check the value directly.

Verified that this fixes my pairing issues with the BTech UV-Pro.  I'm guessing similar issues exist for the other brands of the same radio, but I don't have those to check against.