Update et-vr-n76
When the jq (json query) command executes, it returns 0 if there is a .bluetooth.mac node in the json (which there is, and should be).  This means we can't use $? to see if there is a value in the node or not.  We need to check the value directly.

Verified that this fixes my pairing issues with the BTech UV-Pro in another commit.  I don't have a vr-n76 but the code looks like it isn't radio-specific.